backend/requirements.txt
fastapi
uvicorn
websockets
SQLAlchemy
sqlmodel
python-jose[cryptography]
passlib[bcrypt]